rename CytoscapeLayout to CytoscapeGraph
