+++ src/components/CytoscapeGraph/__tests__/CytoscapeGraph.test.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
 
-import ReactCytoscapeConnected from '../ReactCytoscape';
 import { CytoscapeGraph } from '../CytoscapeGraph';
 import * as GRAPH_DATA from '../../../services/__mockData__/getGraphElements';
 import { Duration, Layout } from '../../../types/GraphFilter';
