--- src/components/CytoscapeLayout/__stories__/bookinfo.json
@@ -1,145 +1,0 @@
-{
-  "nodes": [
-    {
-      "data": {
-        "id": "n2",
-        "text": "details v1",
-        "style": "solid",
-        "service": "details.bookinfo.svc.cluster.local",
-        "version": "v1"
-      }
-    },
-    {
-      "data": {
-        "id": "n1",
-        "text": "productpage v1",
-        "style": "solid",
-        "service": "productpage.bookinfo.svc.cluster.local",
-        "version": "v1",
-        "rate": "36.327",
-        "rate4XX": "36.308"
-      }
-    },
-    {
-      "data": {
-        "id": "n4",
-        "text": "ratings v1",
-        "style": "solid",
-        "service": "ratings.bookinfo.svc.cluster.local",
-        "version": "v1"
-      }
-    },
-    {
-      "data": {
-        "id": "n7",
-        "text": "reviews",
-        "style": "solid",
-        "service": "reviews.bookinfo.svc.cluster.local",
-        "groupBy": "version"
-      }
-    },
-    {
-      "data": {
-        "id": "n5",
-        "text": "reviews v1",
-        "style": "dotted",
-        "parent": "n7",
-        "service": "reviews.bookinfo.svc.cluster.local",
-        "version": "v1"
-      }
-    },
-    {
-      "data": {
-        "id": "n6",
-        "text": "reviews v2",
-        "style": "dotted",
-        "parent": "n7",
-        "service": "reviews.bookinfo.svc.cluster.local",
-        "version": "v2"
-      }
-    },
-    {
-      "data": {
-        "id": "n3",
-        "text": "reviews v3",
-        "style": "solid",
-        "parent": "n7",
-        "service": "reviews.bookinfo.svc.cluster.local",
-        "version": "v3"
-      }
-    },
-    {
-      "data": {
-        "id": "n0",
-        "text": "unknown",
-        "style": "solid",
-        "service": "unknown",
-        "version": "unknown"
-      }
-    }
-  ],
-  "edges": [
-    {
-      "data": {
-        "id": "e0",
-        "source": "n0",
-        "target": "n1",
-        "text": "36.33 99.95%",
-        "color": "red",
-        "style": "solid",
-        "rate": "36.327",
-        "rate4XX": "36.308"
-      }
-    },
-    {
-      "data": {
-        "id": "e1",
-        "source": "n1",
-        "target": "n2",
-        "text": "",
-        "color": "black",
-        "style": "solid"
-      }
-    },
-    {
-      "data": {
-        "id": "e2",
-        "source": "n1",
-        "target": "n3",
-        "text": "",
-        "color": "black",
-        "style": "solid"
-      }
-    },
-    {
-      "data": {
-        "id": "e4",
-        "source": "n1",
-        "target": "n5",
-        "text": "",
-        "color": "black",
-        "style": "dotted"
-      }
-    },
-    {
-      "data": {
-        "id": "e5",
-        "source": "n1",
-        "target": "n6",
-        "text": "",
-        "color": "black",
-        "style": "dotted"
-      }
-    },
-    {
-      "data": {
-        "id": "e3",
-        "source": "n3",
-        "target": "n4",
-        "text": "",
-        "color": "black",
-        "style": "solid"
-      }
-    }
-  ]
-}
